core: fixes to import/OvfDataUpdater issues
The following patch fixes a case in which we attempt to insert a record
to vm_ovf_generations table when we already have a record for the same
guid, which will cause us to pk violation error.
This issue currently occurs in import vm/template flows when we attempt
to use
the same id as exported and the record for that key in
vm_ovf_generations table hasn't been deleted yet due to failure
during the import (compensation doesn't remove record from the ovf
table) or because the vm/template has been removed from the
DB, but not from the storage side by ovf data updater.

Change-Id: I272b39f99d582462c0d1ea4cf268edd7e5c42b6d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=890951
Signed-off-by: Liron Aravot <laravot@redhat.com>