core: DB fix skipped insertions to schema_version
When running create_db.sh scripts 03_00_0420 and beyond are executed but
not inserted to the schema_version table.
This is caused to a race since execute_comamnd is running in a loop and
creates a temporary file in each iteration.
Solution is to gather all inserts in one file and execute them at once.

Change-Id: I285434b566a17e1c1b051469b72d1470bca2109b
Signed-off-by: Eli Mesika <emesika@redhat.com>