restapi: adding host specifying cluster by-name is not supported(#890329)
Using cli, when trying to add a host to a cluster the cluster
name is not supported as a way to indicate which cluster the
host should be added to.

This patch exposes the already existing functionality in the
api code to add host to a cluster by cluster name.

Change-Id: I6b8268b016d83f3a963774b097f78f980503e8dd
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=890329
Signed-off-by: Ravi Nori <rnori@redhat.com>