core: Replace ImagesHandler#isVmDown method
Replace ImagesHandler#isVmDown method with VM#isDown method.

ImagesHander is not the place for utility method that checks VM status.
VM status is a property of VM, thus VM instance should be able to
express whether it's down or not, passing a VM instance to another
class in order to check its status doesn't make sense.

Change-Id: I3f15c8b9a307d5ecb239750dbdbac3faea7b802a
Signed-off-by: Arik Hadas <ahadas@redhat.com>