core: Move status check methods from VM to VMStatus
VM class contained methods which only check the value of the status of
the VM, to see if it's one of several values that reflect some state of
the VM. these methods should be places in VMStatus class (without being
static methods) - instance of VMStatus should be able to return whether
it reflects certain state of the VM, instead of using the previous methods.

While moving those methods to VMStatus, they were renamed such that
they'll be less confusing (isStatusUp is not a good name for a method that
returns true even if the status is not UP for example).

The method VM#isStatusUpOrPausedOrSuspended is removed as it wasn't used
anywhere.

The method VM#isStatusUp was renamed to VM#isRunning, to describe itself
better.

Change-Id: I82b4f109751773ff2d46407ccea26ea0be2661f5
Signed-off-by: Arik Hadas <ahadas@redhat.com>