packaging: Use temporary modules directory
The application server creates .jar.index files in the modules directory
in order to speed up class scanning and class loading. When the
application server is updated those files are not refreshed unless the
application server is started with an user that has permission to write
to the /usr/share/jbossas/modules directory. The engine doesn't have
those permissions, so it can end up using out of date indexes, and this
can generate class loading this problems. To address that issue this
patch changes the engine start script so that it creates a temporary
copy of the modules directory, containing links to the .xml and .jar
files and excluding the .jar.index files.

Change-Id: I4f27390a232df42b477dc0e9abe7909e10d9abb1
Bug-Url: https://bugzilla.redhat.com/889217
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>