webadmin: improve behavior of bookmark list (2/2)
This patch series fixes some problems with the behavior of the bookmark
list.  This patch fixes re-selection of already-selected bookmarks,
which was broken by gwt issue 6310 (which causes setSelectedItem to not
behave as expected when a SingleSelectionModel's keyboardSelectionPolicy
is BOUND_TO_SELECTION).

To work around the problem, we use a different keyboardSelectionPolicy
that isn't affected by the bug, and handle the keypress events manually.

Change-Id: I25a9e721d8b7ccd982708eadf579e9eb229dedd5
Bug-Url: https://bugzilla.redhat.com/880499
Signed-off-by: Greg Padgett <gpadgett@redhat.com>