core: improve logging
This patch introduces several logging improvements, mostly around
backend classes initialized on system startup.
1. Remove new Date() from logging statements - the timestamp is already
part of the log on each line, no need to write it twice
2. Wherever there are no parameters in log statement, just a string, call to log.info
instead of infoFormat which performs several unneeded expensive transformations on
strings.
3. Move initialization log statements to the classes which are
initialized, instead of putting them in the calling class.
4. Make the log statements more clear - instead of just mentioning
hardcoded name of class in the log, specify whether this is start/end of its
initialization so whoever reads the log can understand at which phase
the system is or at which point it failed/stuck.
5. Make logged class name extracted generically instead of hardcoding it

Change-Id: Iaea9c16e9913cdee4eef88de920e87bde8bc6b5d
Signed-off-by: Alissa Bonas <abonas@redhat.com>