packaging: Add support for SSLv3
With the current configuration when we don't use Apache as a proxy we
only support TLSv1. This patch makes the set of supported protocols
configurable.

In order to change the configuration the user can change the
ENGINE_HTTPS_PROTOCOLS parameter in /etc/sysconfig/ovirt-engine. The
default is the following:

  ENGINE_HTTPS_PROTOCOLS=SSLv3,TLSv1,TLSv1.1,TLSv1.2

Note that SSLv2 is not implemented in OpenJDK, so it can't be enabled.

Change-Id: If5f98ede332d717468bf4f55a03e38956fb8e875
Bug-Url: https://bugzilla.redhat.com/893979
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>