engine: different log message for graceful shutdown
Modifies the notifier service to write a different logmessage if the
engine service was shut down by administrator by checking if the pid
file exists. If the pid file exists, the engine crashed, the proper
shutdown script removes the pid file. The availability of the engine
will be logged in any way, the logic only decides the message and the
level.

Change-Id: I409f335fd98959aab4bc1633191a853777b050fa
Signed-off-by: Laszlo Hornyak <lhornyak@redhat.com>