packaging: Disable INFO for malformed cookies
Tomcat (and thus JBoss Web) complains when it finds certain unexpected
characters in cookies, in particular when it finds square brackets. We
don't use such cookies in the applications, but some tracking cookies
have this problem, and they are very frequent. When a browser running
webadmin or user portal sends those cookies it can easily flood the
server.log file with this INFO messages. This patch changes the logging
configuration so that those INFO messages are discarded.

Change-Id: I52f89da503805f94582c955af9a29f6b57b80685
Bug-Url: https://bugzilla.redhat.com/890165
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>