webadmin: disk search overridden by default upon tab reveal
If a search was done on disks and revealing the tab caused the radio
button for the disk view to switch (from All or Luns to Images), the
default search would run instead of the one entered.

The function to filter the search based on the selected image type was
overwriting the entire search string in some instances.  It will now
normalize the input before filtering so ensure that the current search
string can be filtered according to image type.

Change-Id: Iebf63fc5e6343819062959cd85219b1a965daf95
Bug-Url: https://bugzilla.redhat.com/893259
Signed-off-by: Greg Padgett <gpadgett@redhat.com>