engine: Configure management network at host activation
If the engine detects a host which has no management network (either
due to new installation or misconfiguration), the host activation
command will attempt to configure a management network on top of that
interface.

The network configuration on the host will relies on the same
information as reported by VDSM for the host (boot protocol, IP
address and subnet mask).

Change-Id: Ia82e594b888b84b3aa079cc9aa76fa0dddc59f1d
Bug-Url: http://bugzilla.redhat.com/891056
Signed-off-by: Moti Asayag <masayag@redhat.com>