core: Remove VMStatus#isGuestUp method
The method VMStatus#isGuestUp is removed.

The method was called from one place in VdsUpdateRunTimeInfo to
check whether the status of a VM that was in MigrationFrom status
changed to Up/PoweringDown/PoweredDown/PoweringUp in the source host,
which means that the migration failed and needs to be re-run.

The above call was replaced with a check that the VM is not in
MigrationFrom status and a call to VMStatus#isRunning - in those cases
(which include WaitForLaunch + RebootInProgress in addition to the
statuses mentioned above) we want the migration process to re-run.

Change-Id: Iab3535e3bd6abaf61b244ac456bb93eb0218174e
Signed-off-by: Arik Hadas <ahadas@redhat.com>