packaging: Allow passwordless database connection
Currently if no database password is provided in the configuration the
engine will not be able to connect to the database, even if the database
doesn't require a password (it is configured with "trust" mode). This
patch changes the start script so that it will allow this passwordless
connections as this is what we currently do in development environments.

Change-Id: I1c51485720aa7e3289a067e2677239a44d689bf1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>