core: RemoveImageCommand - avoid image staying LOCKED, revert to ILLEGAL
When running remove image command, a DeleteImageGroup vdsm task is being
initiated by the engine. In case of failure or engine crash, the engine
can't tell whether the task has been initiated on vdsm side or not -
therefore the image status should be set to ILLEGAL, in the current
situation the image just stayed LOCKED.

NOTE: the lockImage() with no compensation was left in BaseImagesCommand
as it's used by commands that are being executed within transaction,
so until they will changed to be non transactive, there's no need to
start new transaction within their execution.

Change-Id: Ib5ec91a5f8b8495b8cfd11b19016fee5993644ba
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=840525
Signed-off-by: Liron Aravot <laravot@redhat.com>