core: VM is down check when adding template
AddTemplateCommand explicitly checks if the VM is down in the
canDoAction, so there is no need to recheck it again by passing true to
the relevant argument of PerformImagesChecks.

This patch contains the proposed fix, a test for it, and a minor
refactor required to facilitate the test.

Note: This patch is part of a series intended to remove the VM status
check from ImagesHandler.

Change-Id: Ifec2d9677765502ca6876b31835340d3c4fad5ae
Signed-off-by: Allon Mureinik <amureini@redhat.com>