core: Consolidate VM down check when removing disk
Both removing an image and a LUN disk had to check that the VM is down,
and this check was done in two separate places.
This patch consolidates this treatment to one function and stops using
ImagesHandler to perform this check for images.

Note: This patch is part of a series intended to remove the VM status
check from ImagesHandler.

Change-Id: Ia1a23e1f8d129bf19c1d217c7b9201dc17a2d2c0
Signed-off-by: Allon Mureinik <amureini@redhat.com>