core: RemoveSnapshot vmDown validation
Added an explicit test for VM.isDown in
RemoveSnapshotCommand.canDoAction() instead of using
ImagesHandler.PerformImagesChecks to do it.

In order to do this neatly, such a method was added to VmValidator,
which is the standard way of writing canDoActions.

Note: This patch is part of a series intended to remove the VM status
check from ImagesHandler.

Change-Id: Ifd281aa6778a3249162517499b5f0f48b04ed6b6
Signed-off-by: Allon Mureinik <amureini@redhat.com>