core: TrybackToAllSnapshots vmDown validation
Added an explicit test for VM.isDown in
TrybackToAllSnapshotsOfVmCommand.canDoAction() instead of using
ImagesHandler.PerformImagesChecks to do it.

This patch contains the aforementioned code change, a test for it, and a
minor refactor required to facilitate the test.

Note: This patch is part of a series intended to remove the VM status
check from ImagesHandler.

Change-Id: I5f8fa014e73e6f10693edd1e4ccff4656d50053b
Signed-off-by: Allon Mureinik <amureini@redhat.com>