engine: Removing using of latch all around a code
In engine code has requirements to run couple of simentenious tasks and stuck a main
thread untill they will be finished.
A previous solution was to use a latch, which required to write a same code and keep some
logic for increasing and reducing latch count.
A new solution is to provide a generic way for such problem and to use a standard java api for
running couple of parallel tasks and waits untill all of them will be finished

Change-Id: I34ffc04f7abf57f3d27322a5615f89d07dee16ef
Signed-off-by: Michael Kublin <mkublin@redhat.com>