userportal,webadmin: console error message translation ignored
When the SPICE or RDP console returns an error code, it was ignored.

Fixed by creating a common ConsoleModelErrorEventListener which listens on the
error events and if one occures, shows a popup dialog with the message. It
also translates the error code to human readable form.

This listener is than used in both userportal and webadmin for both RDP and
SPICE

This patch also deletes the unused ResourceManager and Assembly classes which
where not used anymore and had only empty implementations.

Change-Id: Iaa0b0d144dcaafffba8911c0cfa41670e179a7c5
Bug-Url: https://bugzilla.redhat.com/888774
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>