webadmin: Status label is not update on IE
On dragStart we setData on the dragDropEvent.
Later we use this data as a key to the dragged object.

On dragenter, dragleave, dragover we update the status label.
We use the getData to find out the dragged element.
On firefox it works fine, but on most of the browsers drag, dragenter,
dragleave, dragover and dragend don't have access to event's data.
It is part of the DnD specification itself and firefox is the exception
here.

So, to workaround it- the dragged element key is saved as a static member
of NetworkItemPanel class.

Change-Id: Id72896117ce2b26f99607f3d34a55642fbe548bd
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>