webadmin: support IE10
Running an application in IE9 does not necessarily mean you are running IE9
standards mode. IE9 has many modes that can be defined in the page head
tag ("Document Mode").

GWT IE permutations work best with each version of "standards" mode.
Mixing modes, say browser mode = 7 and document mode = 9, is not
recommended and the behavior is undefined.
Gwt's recommendation is "try to keep browser modes and document modes
the same. If you must use mixed mode, be aware that you may run into
issues that are still not supported.
The exception is if you are emulating an older browser when you still do
not support the new version, for instance, you emulate IE7 (EmulateIE7) on
IE9."

It means adding this line-
<meta http-equiv="X-UA-Compatible" content="IE=9">
(tells Internet Explorer to display a webpage in IE9 mode, if possible)
can cause-
1. Problems when using browser with older the IE9 version.
2. Forces IE10 to behave as IE9.

Because of the first cause this line was removed and not
changed to IE=10.

Change-Id: I3ebd720dec57b890b8119ff39efcfb50d34da9e2
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>