gluster: Fetch brick statuses only for UP volumes
Brick statuses are fetched by executing the gluster cli command
'gluster volume status' which works only on UP volumes. Hence
we should not try to fetch and update brick statuses for volumes
that are currently DOWN.

Also made sure that whenever a volume status changes, we update
the status of it's bricks as well.

Introduced more specific audit log types when we detect that a
volume was started/stopped from the gluster cli.

Change-Id: If7f28948036993423d5baad2f4af41875670a18a
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>