gluster: Wrapper method for updating volume status
Introduced a new wrapper method GlusterUtils#updateVolumeStatus() for
updating status of a volume. This internally updates status of all the
bricks of that volume as well.

Modified the start/stop gluster volume commands to make use of this new
method. Removed the method GlusterVolumeCommandBase#updateBrickStatus()
as it is not required any more.

Change-Id: Ia5ae4f24e8c70781d7f93f19c935e136ba8b702b
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>