webadmin, userportal: Don't require path tweak in Eclipse
Currently we need to do some adjustments to the source paths after
importing the projects in Eclipse in order to add the directories
contained generated source. This patch uses the build helper maven
plugin to add those directories, so that Eclipse will import the
projects without the need to manually tweak the paths.

We also need to avoid copying and filtering all the resources to the
generated resources directory, as this generates a validate-build loop
in eclipse.  We can copy and filter only the .gwt.xml files, as this is
what we really need.

Change-Id: Ic48b86f266e287d95872c4048e1d0f6e4f58dd8f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>