restapi: Don't require path tweak in Eclipse
Currently we need to do some adjustments to the source paths after
importing the projects in Eclipse in order to add the directories
contained generated source. This patch uses the build helper maven
plugin to add those directories, so that Eclipse will import the
projects without the need to manually tweak the paths.

Change-Id: Ia55175c60cbef554165e3bae5ffff47d36eb3a18
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>