packaging: Change to / before starting upgrade
If the upgrade runs for a directory that is deleted by the upgrade
itself (anything under /usr/share/rhevm, for example) some tools will
(the JVM, for example) fail because they don't find their current
working directory. In order to prevent that this patch modifies the
upgrade script so that it changes to the root directory before doing
anything.

Change-Id: I5bca7f93fe29e97d17dcc668a0d73306600764cb
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>