core: RemoveVm preview validation
In RemoveVm, moved the validation that the VM is not
in preview from ImagesHandler to SnapshotValidator, which is a more
logical place for it.

Note: VM preview validation is done only if getForce()==true. This is
done to keep consistency with the current behavior, which has been in
place since this git repository was established. Whether or not it is
correct is a different question, which warrants inspection in an
additional patch.

Note: This patch is part of a series of patches aimed at removing the
preview validation from ImagesHandler altogether.

Change-Id: If49d2094589c639736649ee2eadcd45df9872826
Signed-off-by: Allon Mureinik <amureini@redhat.com>