core: RemoveDisk preview validation
In RemoveDisk, moved the validation that the VM is not
in preview from ImagesHandler to SnapshotValidator, which is a more
logical place for it.

This patch does not only clean up the code, it also solves a subtle bug
in RemoveDisk's flow.
Before this patch, the VM's preview status was checked only for plugged
image disks. However, since removing a disk needs to update all of the
VM's OVFs and remove it from them, it would have performed incorrectly
if used on a VM during preview mode.

Note: This patch is part of a series of patches aimed at removing the
preview validation from ImagesHandler altogether.

Change-Id: Iefbea2785b3637ea5df3953d6e8f1a24765742a0
Signed-off-by: Allon Mureinik <amureini@redhat.com>