core: RemoveSnapshot preview validation
In RemoveSnapshot, moved the validation that the VM is not
in preview from ImagesHandler to SnapshotValidator, which is a more
logical place for it.

This patch contains the aforementioned change in RemoveSnapshot's code,
and the required adjustments to the mocking done in
RemoveSnapshotCommandTest to allow it to continue functioning.

Note: This patch is part of a series of patches aimed at removing the
preview validation from ImagesHandler altogether.

Change-Id: Ib42871fecd4738a77c657f07a464c8dd68dee592
Signed-off-by: Allon Mureinik <amureini@redhat.com>