core: VM ImageLocked status check in ImagesHandler
Since we now have statuses per disk, checking a VM's status for
ImageLocked seems unrelated to ImagesHandler.
In fact, this has become a more "logical" locked status, not directly
associated to the VM's images.

The relevant code was removed from ImagesHandler and moved to
VmValidator, where it belongs.

Note that most places that required this validation already validated
first that the VM was in fact down, so adding an additional status check
there was pointless.
In the places where the additional check was warranted, care was taken
to make the VM's status check /AFTER/ the disks', so that the most
informative error message (containing the aliases of the locked disks)
could be produced.

To seal the deal, the API of the relevant methods was changed to use VM
ID instead of a VM, to show it is no longer needed.

Change-Id: Iacfcb81df64b778b2093c7c84a8582cfd59be29c
Signed-off-by: Allon Mureinik <amureini@redhat.com>