core: VM ImageIllegal status check in ImagesHandler
Since we now have statuses per disk, checking a VM's status for
ImageInvalid seems unrelated to ImagesHandler.
In fact, this has become a more "logical" illegal status, not directly
associated to the VM's images.

The relevant code was removed from ImagesHandler and moved to
VmValidator, where it belongs.

Note that only place that used this validation (i.e., checkIsValid=true
&& checkImagesIllegal=true) and did not explicitly validate the VM's
state beforhand (regardless) is CreateAllSnapshotFromVmCommad.

To seal the deal, the API of the relevant methods was changed to use VM
ID instead of a VM, to show it is no longer needed.

Change-Id: I0578a91929689a95845b51a6cd5b1f42ef497c91
Signed-off-by: Allon Mureinik <amureini@redhat.com>