core: cancel migration better error handling
Cancelling migration could fail because the VM already migrated to the
destination host or the VM might crash on the src host.

Those conditions results error return 1, "Virtual machine does not
exist" from the host which is confusing. Instead the user is guided to
try again and try to track the VM events logs.

Change-Id: I546cf86c1c3eeb062b9558d3c82e25035c82efbb
Bug-Url: https://bugzilla.redhat.com/878778
Signed-off-by: Roy Golan <rgolan@redhat.com>