core: ImportVm - avoid early run of end method
When creating tasks for the imported vm images, the parameters entity id
was set to the disk id - this caused that the end method of import vm
command will be called before the tasks on all disks were finished -
causing to the vm and it's disks appear as OK while there not all tasks
has finished running which can cause to hazardous issues and to the end
method to be called for each disk separately.

NOTE: ImportVmTemplate is handled in a different patch.

Change-Id: I0598698bb85c8359b981b643ee26e7cb777707b9
Signed-off-by: Liron Aravot <laravot@redhat.com>