gluster: Do not log removal of option group
The option "group" is a special case, where glusterfs internally
replaces it with the options defined in corresponding group file.
Hence, every time you set the option group, it eventually gets
replaced with a set of options.

For this reason, it is not appropriate to create an audit log when
GlusterManager detects that the option "group" has been removed.
Hence added a check to avoid the same.

Change-Id: Ic1fc981563239ca2209c7172805d802a9691cce4
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>