engine: Allow skipping management network validation
The management network validation is skipped when refresh capabilities
is being executed as part of the host activation flow. In this flow
there is a change the host wasn't configured to have the management
network, therefore management network is removed from the network
validation and it be performed as part of the setup network sequence
in which the management network will be created on the host.

Change-Id: Idf4bf4affcd901b23e2d4d55ba85763928180fcd
Signed-off-by: Moti Asayag <masayag@redhat.com>