core: import vm -persist disks with correct storage domain
When importing vm, the added disks were persisted with the
export domain as their storage domain rather then the target storage
domain.

note: the correct storage domain is set only for the saved disks and
not to all of it's snapshots - if the operation is successful then the
storage domain will be set to all snapshots from end method of
MoveOrCopyImageGroup, otherwise the disk will be removed.

Change-Id: I43280c97f2e444720044966eead7276999bdce1a
Signed-off-by: Liron Aravot <laravot@redhat.com>