core: taskcleaner.sh utility outputs extra lines
When using taskcleaner.sh utility to list task or zombie tasks, the
utility outputs extra lines.
The directories in the output are the result of pushd and popd using in
the script to make it runnable from any place by its full name.
When the utility returns no data, it is important that no redundant
output is generated in order to track this correctly.

Change-Id: I5bf28cfe2311449a0bdd68e132622ced57d4a868
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url:https://bugzilla.redhat.com/show_bug.cgi?id=903156