core: Refactor SP validation out of ImagesHandler
Moved the validation on the SP's status from
ImagesHandler.PerformImagesChecks to StoragePoolValidator, where it
belongs.

In all the cases where this validation was performed by an ImagesHandler
call, a call to StoragePoolValidator was added in order to keep
backwards consistency, even though it is probably useless in many
(most?) cases.
E.g., there is little point in checking if the SP is UP if we already
know we have an UP domain on it.
However, these changes are left for future patches in order not to "rock
the boat" too much in a single patch.

Change-Id: I5caa3724fb73c8fa8932c8071f6d27a78d681bb3
Signed-off-by: Allon Mureinik <amureini@redhat.com>