core: revert change If615a9d9
In patch If615a9d9 the execution context of SetNonOperationalVdsCommand
changed to async if it triggered migrations.

This patch change its execution context back to sync. migrations that
will be triggered from SetNonOperationalCommand will be executed as
different jobs (it will be pushed in a different patch), thus resolving
bug 888199 without making the set-non-operational command to be async
(this command shouldn't be async, as the host become non-operational
right after the sync part, independently from the triggered migrations)

Change-Id: I91eb9e84d0ff4ce6a0c8390dc973c3f8822a0c8c
Signed-off-by: Arik Hadas <ahadas@redhat.com>