core: fix monitoring of SetNonOperationalVdsCommand
This patch resolves a bug in which an exception was thrown when a step
for an already completed and cleaned-up job is saved, when
SetNonOperationalVdsCommand is invoked and finished and then a step for
a migration it triggered is about to save.

The solution is to extract the migrations to be monitored as different
job, which reflects the process better on this scenario as the host become
non-oprational independently of the status of the triggered migrations.

Change-Id: I06b08108b1c75fe00b3f4dc6336c61a5e8c95d85
Bug-Url: https://bugzilla.redhat.com/888199
Signed-off-by: Arik Hadas <ahadas@redhat.com>