core: rename remove storage connection command
Rename removeStorageServerConnection command and VdcAction type
to disconnectStorageServerConnection since this command
disconnects it on vdsm side, but does not remove its record from the db.
It's more correct to call it "disconnect" according to what it actually does.
It's also preparation for adding (in different patch) a remove command
that actully both disconnects the storage and removes the storage connection from
the db.
Also, remove a big commented piece of code from AddStorageDomainRM
because it's commented out anyway and just makes the class bigger
because of dead code.

Change-Id: I0b6043810a5c4200593befd668a279ea56ff4693
Signed-off-by: Alissa Bonas <abonas@redhat.com>