core: ImagesHandler.checkDiskImages with multi-SDs
In the current implementation, checkDiskImages takes the first storage
domain for each disk it checks.
This behaviour is obviously problematic for images which exist on
multiple storage domains, as the check could succeed or fail based solely
on the order the domains appear in the disk's data member.

This patch changes the logic to accumulate ALL of the domains of all the
images, for a more meaningful behaviour.

Looking forward, this change will enable using checkDiskImages for
template disks correctly, even though all the current use cases are VM
usecases, which implies a single domain per disk. Thus, this patch
preserves the current behaviour, but allows additional future usecases.

Change-Id: I61b729bdcbd3001d371fb58a6358172b07be94f4
Signed-off-by: Allon Mureinik <amureini@redhat.com>