core: Fix GetAllVmPoolsAttachedToUserQueryTest
GetAllVmPoolsAttachedToUserQueryTest was broken by change-id
I59346bc9f1cf6d948c4401ddbd7bb365c67b5b05 which changed the query from a
user query to an admin query.

This patch removes the parts of the test that are only relevant to user
queries (i.e., permission filtering) so that it will only check the
expected behavior.

Change-Id: I5883d47c4ca074c83913449ecd80972139a7dcf2
Signed-off-by: Allon Mureinik <amureini@redhat.com>