core: Fix empty quota validation in edit disk
When a disk with no quota is edited and the quota is not setted by the
user, no quota validation takes place. The operation succeeds.

This error is due to wrong logic in UpdateVmDiskCommand.

Logic was change so that only if quota is non-empty and was not changed
by user, it will not be validated.

Change-Id: I697171e9c47499048d5a735c0f0e5a253b41892a
Bug-Url: https://bugzilla.redhat.com/895981
Signed-off-by: Ofri Masad <omasad@redhat.com>