core: handle error when iso domain in maintenance
The bug started from code inserted in change I9d24c94c.
Currently, when iso domain cannot be accessed for any reason, it's considered a
problematic domain and is mentioned in the event log.
This patch adds a check that if the domain is not active,
it is not attempted to be refreshed, hence not adding in event log
the error (because it's expected it cannot be accessed)

Change-Id: Ide53f42c087ffbc9240cccc033e809ff8117ec8c
Bug-Url: https://bugzilla.redhat.com/902246
Signed-off-by: Alissa Bonas <abonas@redhat.com>