webadmin: Add default description to multiple action result
Each result on multiple action result list can have multiple canDo action
messages.
Today all the canDo action messages are displayed as a bulk and the user
can't understand which message belongs to which action.

To correlate between the action and the message a description should be set
on the result.
For example if the user is trying to remove multiple vms. The description
of each result should be the vm name.

The errors are displayed in the following way-
description1:
  - msg1
  - msg2
description2:
  - msg1

The problem is that not all the commands set description on the returned
value.

This patch checks if there is description, if not, a default one is set-
"Action #num:"

Change-Id: I25495eedf3ec61e11eb3c2bb075833fbcc8f9537
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>