webadmin: Fix priority options not presented bug
When all hosts are down or in maintenance mode the priority options in
the "New Server" pop-up window, "High Availability" tab are not
presented.

The priority model is initialized only when a Template is selected.

Added a new initialization point for the priority model in the
initialization of the pop-up window.

Change-Id: I0c66119043625638399972bde95d0ab068d373fe
Bug-Url: https://bugzilla.redhat.com/903726
Signed-off-by: Ofri Masad <omasad@redhat.com>