core: Enable remove quota when quota is not enforced
In current implementation there is no way to remove all quota from the
DC once a quota was assigned to some entities.

The CanDoAction logic of removing quota was changed so that when a DC is
in 'Disabled' Quota enforcement mode a quota could be removed
(regardless of it been assigned to some entities).

Change-Id: I95ff993b0db21a17e62e3258338bd524c5c8b957
Bug-Url: https://bugzilla.redhat.com/895237
Signed-off-by: Ofri Masad <omasad@redhat.com>