Hosts from another data center displayed when...
Hosts from another data center displayed when clicked on Hosts in tree view for cluster names with a single difference

The way the ESCAPE character is used was changed between 8.4 and 9.x
The '_' character is considered as regexp '?' character in the case
ilike is used as the condition operator, therefore, we are forced to
sign it as a normal character that is part of the string
However, it seems that PG 8.4 forces us to use double ESCAPE character
for that while it will not find any matching with one ESCAPE character
and PG 9.1 is exactly the opposite , one ESCAPE character works while
double ESCAPE characters will bring no data.

The patch adds a new configuration variable PgMajorRelease that is
computed when the 0000_config.sql script is running.
The value of this variable is used to determine how to skip the special
character '_' if found in the value string.

Change-Id: I808fe02553b5554b76f08b43a30fb0ffd7d99973
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=874660